requirements.txt
# Data Processing & Analytics
# Note: Requires Python 3.11+

pandas>=2.0.0
numpy>=1.24.0
polars>=0.19.0
duckdb>=0.9.0
pyarrow>=12.0.0

# Backtesting & Technical Analysis
# vectorbt>=0.25.0  # Optional - may have compatibility issues with Python 3.14
# Note: VectorBT has known issues with Python 3.14 (sklearn/joblib compatibility)
# Install separately if needed: pip install vectorbt
# Consider using Python 3.11 or 3.12 for full VectorBT compatibility
# pandas-ta>=0.3.14b0  # Temporarily disabled - incompatible with Python 3.14
# Note: Using custom indicator implementation instead

# Async HTTP & Data Ingestion
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# UI Framework
streamlit>=1.28.0
# pywebview>=4.4.0  # Optional - may have compatibility issues, install separately if needed
plotly>=5.17.0

# Telegram Bot
python-telegram-bot>=20.7

# Configuration & Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Logging & Monitoring
loguru>=0.7.2
